+++ src/State/AbstractHostState.php
@@ -37,7 +37,10 @@ abstract class AbstractHostState implements State
                 // 2.1. If buffer is the empty string, validation error, return failure.
                 if ($context->buffer->isEmpty()) {
                     // Validation error. Return failure.
-                    $context->logger?->warning('unexpected-port-without-host');
+                    $context->logger?->warning('unexpected-port-without-host', [
+                        'input'  => (string) $context->input,
+                        'column' => $context->iter->key() + 1,
+                    ]);
 
                     return self::RETURN_FAILURE;
                 }
@@ -82,7 +85,10 @@ abstract class AbstractHostState implements State
                 // 3.1. If url is special and buffer is the empty string, validation error, return failure.
                 if ($context->url->scheme->isSpecial() && $context->buffer->isEmpty()) {
                     // Validation error. Return failure.
-                    $context->logger?->warning('empty-host-special-scheme');
+                    $context->logger?->warning('empty-host-special-scheme', [
+                        'input'  => (string) $context->input,
+                        'column' => $context->iter->key() + 2, // Add 2 since we called ->prev() above
+                    ]);
 
                     return self::RETURN_FAILURE;
                 }
@@ -95,7 +101,10 @@ abstract class AbstractHostState implements State
                     && ($context->url->includesCredentials() || $context->url->port !== null)
                 ) {
                     // Validation error.
-                    $context->logger?->notice('host-invalid');
+                    $context->logger?->notice('host-invalid', [
+                        'input'  => (string) $context->input,
+                        'column' => $context->iter->key() + 2, // Add 2 since we called ->prev() above
+                    ]);
 
                     return self::RETURN_BREAK;
                 }