+++ src/State/RelativeSlashState.php
@@ -22,7 +22,10 @@ class RelativeSlashState implements State
             // 1.1. If c is U+005C (\), validation error.
             if ($codePoint === '\\') {
                 // Validation error.
-                $context->logger?->notice('unexpected-reverse-solidus');
+                $context->logger?->notice('unexpected-reverse-solidus', [
+                    'input'  => (string) $context->input,
+                    'column' => $context->iter->key() + 1,
+                ]);
             }
 
             // 1.2. Set state to special authority ignore slashes state.